split check command file read host effects
